1.extract class 2. EncryptDecryptUtils -> EncryptUtil and DecryptUtil 3. Since the encryption and decryption processes are different, I have re-created the other two classes.
1. extract method
2. DecryptUtil - decryptFile method
3. Toasts are overlap a lot. This part is extracted.
    For the understandability of decryptFile method, methods were extracted for each case.
